Change the way of the key input
To be simple.

Previously, the key events such as 'keydown' and 'keyup' were listened by
a button object which has a focus, So, Whether the MZ-700 emulator
processes the events is depend on which object has a focus.

This time, it was changed to receive the events by window object, and
whether the events are handled by the emulator is depend on the internal
flag in the MZ-700 object.

If the screen area of the emulator is clicked, the events are handled by
the emulator. Otherwise, those are handled by the browsers.