Fix the bugs in LOAD command of NEWMON7
The CMT reading process `rdata` of MZ_DataRecorder was revised.

Some fixed bugs at fromBytes of MZ_Tape are:
On creating pseudo magnetic signal stream for CMT,
bits in a byte must be recorded from MSB to LSB,
and a byte-order of the 16-bit checksum must be big endian.

a test script for the MZ_Tape is added.