v0.9.17 - Fix and close the issues #76 and #84
For issue #76, Handle scroll event to update the top-left address.
The dump list could not update its top left address when the list was
scrolled. And the Vscrlist plug-in had not fired the event.

For issue #84, Revive the bottom border of the dump list header
The CSS was not following the previous changes.