Refactoring
The Web server is changed to the 'http-server' from 'express' and the
server's root directory is moved to the parent directory of this module.

The source files of Cli-commands are moved below 'lib'.

And also, those which relating modules, source files, and directories are
changed.

The binaries in a directory 'tools' is moved below 'bin'.