Add unit tests for MZ700_memory.changeBlock1_VRAM/DRAM
SUGI_ML.MZT stops at 89C4H.
At the time, The mnemonic is 'LD A,(DE)' and register DE is D000H.

When the error above was caught, same error is thrown by reading D000
by the memory dump widget.

So, I assume that the bug is related to the memory bank changing feature.

This commit is relating to the issue #115.