Improve the speed of the emulator which using WebSocket interface
It had been taking times at the MMIO write access for PCG-700.
Moving the PCG-700 implementation to the server side improved the speed.

About the updating screen, drawing characters are done on server side.
But, the whole canvas on browser is redrawing with the image data sent
from server side.